requirements.txt
cffi==1.15.1
Cython==3.0.2
easydict==1.10
einops==0.3.0
h5py==3.9.0
imageio==2.9.0
ipython==8.15.0
numpy==1.21.6
opencv_python==4.1.2.30
opencv_python_headless==4.7.0.72
Pillow==9.3.0
Pillow==10.0.0
pycocotools==2.0.7
PyYAML==6.0
PyYAML==6.0.1
scipy==1.9.1
setuptools==65.6.3
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
tensorflow==2.13.0
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.65.0
